+++ lib/bluetooth/devices/zwift/zwift_click.dart
@@ -3,16 +3,17 @@ import 'package:swift_control/bluetooth/devices/zwift/protocol/zwift.pb.dart';
 import 'package:swift_control/bluetooth/devices/zwift/zwift_device.dart';
 import 'package:swift_control/utils/keymap/buttons.dart';
 
+import 'constants.dart';
+
 class ZwiftClick extends ZwiftDevice {
-  ZwiftClick(super.scanResult)
-    : super(availableButtons: [ControllerButton.shiftUpRight, ControllerButton.shiftDownLeft]);
+  ZwiftClick(super.scanResult) : super(availableButtons: [ZwiftButtons.shiftUpRight, ZwiftButtons.shiftDownLeft]);
 
   @override
   List<ControllerButton> processClickNotification(Uint8List message) {
     final status = ClickKeyPadStatus.fromBuffer(message);
     final buttonsClicked = [
-      if (status.buttonPlus == PlayButtonStatus.ON) ControllerButton.shiftUpRight,
-      if (status.buttonMinus == PlayButtonStatus.ON) ControllerButton.shiftDownLeft,
+      if (status.buttonPlus == PlayButtonStatus.ON) ZwiftButtons.shiftUpRight,
+      if (status.buttonMinus == PlayButtonStatus.ON) ZwiftButtons.shiftDownLeft,
     ];
     return buttonsClicked;
   }