+++ lib/bluetooth/devices/zwift/zwift_ride.dart
@@ -21,24 +21,24 @@ class ZwiftRide extends ZwiftDevice {
   ZwiftRide(super.scanResult, {super.isBeta})
     : super(
         availableButtons: [
-          ControllerButton.navigationLeft,
-          ControllerButton.navigationRight,
-          ControllerButton.navigationUp,
-          ControllerButton.navigationDown,
-          ControllerButton.a,
-          ControllerButton.b,
-          ControllerButton.y,
-          ControllerButton.z,
-          ControllerButton.shiftUpLeft,
-          ControllerButton.shiftDownLeft,
-          ControllerButton.shiftUpRight,
-          ControllerButton.shiftDownRight,
-          ControllerButton.powerUpLeft,
-          ControllerButton.powerUpRight,
-          ControllerButton.onOffLeft,
-          ControllerButton.onOffRight,
-          ControllerButton.paddleLeft,
-          ControllerButton.paddleRight,
+          ZwiftButtons.navigationLeft,
+          ZwiftButtons.navigationRight,
+          ZwiftButtons.navigationUp,
+          ZwiftButtons.navigationDown,
+          ZwiftButtons.a,
+          ZwiftButtons.b,
+          ZwiftButtons.y,
+          ZwiftButtons.z,
+          ZwiftButtons.shiftUpLeft,
+          ZwiftButtons.shiftDownLeft,
+          ZwiftButtons.shiftUpRight,
+          ZwiftButtons.shiftDownRight,
+          ZwiftButtons.powerUpLeft,
+          ZwiftButtons.powerUpRight,
+          ZwiftButtons.onOffLeft,
+          ZwiftButtons.onOffRight,
+          ZwiftButtons.paddleLeft,
+          ZwiftButtons.paddleRight,
         ],
       );
 
@@ -184,31 +184,24 @@ class ZwiftRide extends ZwiftDevice {
 
     // Process DIGITAL buttons separately
     final buttonsClicked = [
-      if (status.buttonMap & _RideButtonMask.LEFT_BTN.mask == PlayButtonStatus.ON.value)
-        ControllerButton.navigationLeft,
-      if (status.buttonMap & _RideButtonMask.RIGHT_BTN.mask == PlayButtonStatus.ON.value)
-        ControllerButton.navigationRight,
-      if (status.buttonMap & _RideButtonMask.UP_BTN.mask == PlayButtonStatus.ON.value) ControllerButton.navigationUp,
-      if (status.buttonMap & _RideButtonMask.DOWN_BTN.mask == PlayButtonStatus.ON.value)
-        ControllerButton.navigationDown,
-      if (status.buttonMap & _RideButtonMask.A_BTN.mask == PlayButtonStatus.ON.value) ControllerButton.a,
-      if (status.buttonMap & _RideButtonMask.B_BTN.mask == PlayButtonStatus.ON.value) ControllerButton.b,
-      if (status.buttonMap & _RideButtonMask.Y_BTN.mask == PlayButtonStatus.ON.value) ControllerButton.y,
-      if (status.buttonMap & _RideButtonMask.Z_BTN.mask == PlayButtonStatus.ON.value) ControllerButton.z,
-      if (status.buttonMap & _RideButtonMask.SHFT_UP_L_BTN.mask == PlayButtonStatus.ON.value)
-        ControllerButton.shiftUpLeft,
+      if (status.buttonMap & _RideButtonMask.LEFT_BTN.mask == PlayButtonStatus.ON.value) ZwiftButtons.navigationLeft,
+      if (status.buttonMap & _RideButtonMask.RIGHT_BTN.mask == PlayButtonStatus.ON.value) ZwiftButtons.navigationRight,
+      if (status.buttonMap & _RideButtonMask.UP_BTN.mask == PlayButtonStatus.ON.value) ZwiftButtons.navigationUp,
+      if (status.buttonMap & _RideButtonMask.DOWN_BTN.mask == PlayButtonStatus.ON.value) ZwiftButtons.navigationDown,
+      if (status.buttonMap & _RideButtonMask.A_BTN.mask == PlayButtonStatus.ON.value) ZwiftButtons.a,
+      if (status.buttonMap & _RideButtonMask.B_BTN.mask == PlayButtonStatus.ON.value) ZwiftButtons.b,
+      if (status.buttonMap & _RideButtonMask.Y_BTN.mask == PlayButtonStatus.ON.value) ZwiftButtons.y,
+      if (status.buttonMap & _RideButtonMask.Z_BTN.mask == PlayButtonStatus.ON.value) ZwiftButtons.z,
+      if (status.buttonMap & _RideButtonMask.SHFT_UP_L_BTN.mask == PlayButtonStatus.ON.value) ZwiftButtons.shiftUpLeft,
       if (status.buttonMap & _RideButtonMask.SHFT_DN_L_BTN.mask == PlayButtonStatus.ON.value)
-        ControllerButton.shiftDownLeft,
-      if (status.buttonMap & _RideButtonMask.SHFT_UP_R_BTN.mask == PlayButtonStatus.ON.value)
-        ControllerButton.shiftUpRight,
+        ZwiftButtons.shiftDownLeft,
+      if (status.buttonMap & _RideButtonMask.SHFT_UP_R_BTN.mask == PlayButtonStatus.ON.value) ZwiftButtons.shiftUpRight,
       if (status.buttonMap & _RideButtonMask.SHFT_DN_R_BTN.mask == PlayButtonStatus.ON.value)
-        ControllerButton.shiftDownRight,
-      if (status.buttonMap & _RideButtonMask.POWERUP_L_BTN.mask == PlayButtonStatus.ON.value)
-        ControllerButton.powerUpLeft,
-      if (status.buttonMap & _RideButtonMask.POWERUP_R_BTN.mask == PlayButtonStatus.ON.value)
-        ControllerButton.powerUpRight,
-      if (status.buttonMap & _RideButtonMask.ONOFF_L_BTN.mask == PlayButtonStatus.ON.value) ControllerButton.onOffLeft,
-      if (status.buttonMap & _RideButtonMask.ONOFF_R_BTN.mask == PlayButtonStatus.ON.value) ControllerButton.onOffRight,
+        ZwiftButtons.shiftDownRight,
+      if (status.buttonMap & _RideButtonMask.POWERUP_L_BTN.mask == PlayButtonStatus.ON.value) ZwiftButtons.powerUpLeft,
+      if (status.buttonMap & _RideButtonMask.POWERUP_R_BTN.mask == PlayButtonStatus.ON.value) ZwiftButtons.powerUpRight,
+      if (status.buttonMap & _RideButtonMask.ONOFF_L_BTN.mask == PlayButtonStatus.ON.value) ZwiftButtons.onOffLeft,
+      if (status.buttonMap & _RideButtonMask.ONOFF_R_BTN.mask == PlayButtonStatus.ON.value) ZwiftButtons.onOffRight,
     ];
 
     // Process ANALOG inputs separately - now properly separated from digital
@@ -219,8 +212,8 @@ class ZwiftRide extends ZwiftDevice {
         if (paddle.hasLocation() && paddle.hasAnalogValue()) {
           if (paddle.analogValue.abs() >= analogPaddleThreshold) {
             final button = switch (paddle.location.value) {
-              0 => ControllerButton.paddleLeft, // L0 = left paddle
-              1 => ControllerButton.paddleRight, // L1 = right paddle
+              0 => ZwiftButtons.paddleLeft, // L0 = left paddle
+              1 => ZwiftButtons.paddleRight, // L1 = right paddle
               _ => null, // L2, L3 unused
             };