+++ lib/pages/touch_area.dart
@@ -4,26 +4,25 @@ import 'dart:math';
 
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter/scheduler.dart';
 import 'package:flutter/services.dart';
 import 'package:image_picker/image_picker.dart';
-import 'package:keypress_simulator/keypress_simulator.dart';
+import 'package:path_provider/path_provider.dart';
 import 'package:swift_control/main.dart';
+import 'package:swift_control/widgets/button_widget.dart';
 import 'package:swift_control/widgets/keymap_explanation.dart';
 import 'package:swift_control/widgets/menu.dart';
 import 'package:swift_control/widgets/testbed.dart';
 import 'package:window_manager/window_manager.dart';
 
-import '../bluetooth/messages/notification.dart';
 import '../utils/actions/base_actions.dart';
-import '../utils/keymap/apps/custom_app.dart';
-import '../utils/keymap/buttons.dart';
 import '../utils/keymap/keymap.dart';
-import '../widgets/custom_keymap_selector.dart';
 
 final touchAreaSize = 42.0;
 
 class TouchAreaSetupPage extends StatefulWidget {
-  const TouchAreaSetupPage({super.key});
+  final KeyPair keyPair;
+  const TouchAreaSetupPage({super.key, required this.keyPair});
 
   @override
   State<TouchAreaSetupPage> createState() => _TouchAreaSetupPageState();
@@ -31,8 +30,6 @@ class TouchAreaSetupPage extends StatefulWidget {
 
 class _TouchAreaSetupPageState extends State<TouchAreaSetupPage> {
   File? _backgroundImage;
-  late StreamSubscription<BaseNotification> _actionSubscription;
-  ControllerButton? _pressedButton;
   final TransformationController _transformationController = TransformationController();
 
   late Rect _imageRect;
@@ -42,29 +39,37 @@ class _TouchAreaSetupPageState extends State<TouchAreaSetupPage> {
     final result = await picker.pickImage(source: ImageSource.gallery);
     if (result != null) {
       final image = File(result.path);
-
-      // need to decode image to get its size so we can have a percentage mapping
-      final decodedImage = await decodeImageFromList(image.readAsBytesSync());
-      // calculate image rectangle in the current screen, given it's boxfit contain
-      final screenSize = MediaQuery.sizeOf(context);
-      final imageAspectRatio = decodedImage.width / decodedImage.height;
-      final screenAspectRatio = screenSize.width / screenSize.height;
-      if (imageAspectRatio > screenAspectRatio) {
-        // image is wider than screen
-        final width = screenSize.width;
-        final height = width / imageAspectRatio;
-        final top = (screenSize.height - height) / 2;
-        _imageRect = Rect.fromLTWH(0, top, width, height);
-      } else {
-        // image is taller than screen
-        final height = screenSize.height;
-        final width = height * imageAspectRatio;
-        final left = (screenSize.width - width) / 2;
-        _imageRect = Rect.fromLTWH(left, 0, width, height);
-      }
-      _backgroundImage = image;
-      setState(() {});
+      final Directory tempDir = await getTemporaryDirectory();
+      final tempImage = File('${tempDir.path}/${actionHandler.supportedApp?.name ?? 'temp'}_screenshot.png');
+      await image.copy(tempImage.path);
+      _backgroundImage = tempImage;
+      await _calculateBounds();
     }
+  }
+
+  Future<void> _calculateBounds() async {
+    if (_backgroundImage == null) return;
+
+    // need to decode image to get its size so we can have a percentage mapping
+    final decodedImage = await decodeImageFromList(_backgroundImage!.readAsBytesSync());
+    // calculate image rectangle in the current screen, given it's boxfit contain
+    final screenSize = MediaQuery.sizeOf(context);
+    final imageAspectRatio = decodedImage.width / decodedImage.height;
+    final screenAspectRatio = screenSize.width / screenSize.height;
+    if (imageAspectRatio > screenAspectRatio) {
+      // image is wider than screen
+      final width = screenSize.width;
+      final height = width / imageAspectRatio;
+      final top = (screenSize.height - height) / 2;
+      _imageRect = Rect.fromLTWH(0, top, width, height);
+    } else {
+      // image is taller than screen
+      final height = screenSize.height;
+      final width = height * imageAspectRatio;
+      final left = (screenSize.width - width) / 2;
+      _imageRect = Rect.fromLTWH(left, 0, width, height);
+    }
+    setState(() {});
   }
 
   void _saveAndClose() {
@@ -74,7 +79,6 @@ class _TouchAreaSetupPageState extends State<TouchAreaSetupPage> {
   @override
   void dispose() {
     super.dispose();
-    _actionSubscription.cancel();
     // Exit full screen
     SystemChrome.setEnabledSystemUIMode(SystemUiMode.manual, overlays: SystemUiOverlay.values);
     // Reset orientation preferences to allow all orientations
@@ -105,35 +109,16 @@ class _TouchAreaSetupPageState extends State<TouchAreaSetupPage> {
     if (Platform.isWindows || Platform.isLinux || Platform.isMacOS) {
       windowManager.setFullScreen(true);
     }
-    _actionSubscription = connection.actionStream.listen((data) async {
-      if (!mounted) {
-        return;
-      }
-      if (data is ButtonNotification) {
-        _pressedButton = data.buttonsClicked.singleOrNull;
-      }
+    getTemporaryDirectory().then((tempDir) async {
+      final tempImage = File('${tempDir.path}/${actionHandler.supportedApp?.name ?? 'temp'}_screenshot.png');
+      if (tempImage.existsSync()) {
+        _backgroundImage = tempImage;
+        setState(() {});
 
-      if (_pressedButton != null) {
-        if (actionHandler.supportedApp!.keymap.getKeyPair(_pressedButton!) == null) {
-          final KeyPair keyPair;
-          actionHandler.supportedApp!.keymap.keyPairs.add(
-            keyPair = KeyPair(
-              touchPosition: Offset((actionHandler.supportedApp!.keymap.keyPairs.length + 1) * 10, 10),
-              buttons: [_pressedButton!],
-              physicalKey: null,
-              logicalKey: null,
-              isLongPress: false,
-            ),
-          );
-          setState(() {});
-
-          // open menu
-          if (Platform.isMacOS || Platform.isWindows) {
-            await Future.delayed(Duration(milliseconds: 300));
-            await keyPressSimulator.simulateMouseClickDown(keyPair.touchPosition);
-            await keyPressSimulator.simulateMouseClickUp(keyPair.touchPosition);
-          }
-        }
+        // wait a bit until device rotation is done
+        SchedulerBinding.instance.addPostFrameCallback((_) {
+          _calculateBounds();
+        });
       }
     });
   }
@@ -173,93 +158,6 @@ class _TouchAreaSetupPageState extends State<TouchAreaSetupPage> {
       _imageRect.top + relativeY * _imageRect.height - differenceInHeight - iconSize / 2,
     );
 
-    final actions = [
-      if (actionHandler.supportedModes.contains(SupportedMode.keyboard))
-        PopupMenuItem<PhysicalKeyboardKey>(
-          value: null,
-          child: ListTile(
-            leading: Icon(Icons.keyboard_alt_outlined),
-            title: const Text('Simulate Keyboard shortcut'),
-            trailing: keyPair.physicalKey != null ? Checkbox(value: true, onChanged: null) : null,
-          ),
-          onTap: () async {
-            await showDialog<void>(
-              context: context,
-              barrierDismissible: false, // enable Escape key
-              builder: (c) =>
-                  HotKeyListenerDialog(customApp: actionHandler.supportedApp! as CustomApp, keyPair: keyPair),
-            );
-            setState(() {});
-          },
-        ),
-      if (actionHandler.supportedModes.contains(SupportedMode.touch))
-        PopupMenuItem<PhysicalKeyboardKey>(
-          value: null,
-          child: ListTile(
-            title: const Text('Simulate Touch'),
-            leading: Icon(Icons.touch_app_outlined),
-            trailing: keyPair.physicalKey == null && keyPair.touchPosition != Offset.zero
-                ? Checkbox(value: true, onChanged: null)
-                : null,
-          ),
-          onTap: () {
-            keyPair.physicalKey = null;
-            keyPair.logicalKey = null;
-            setState(() {});
-          },
-        ),
-
-      if (actionHandler.supportedModes.contains(SupportedMode.media))
-        PopupMenuItem<PhysicalKeyboardKey>(
-          child: PopupMenuButton<PhysicalKeyboardKey>(
-            padding: EdgeInsets.zero,
-            itemBuilder: (context) => [
-              PopupMenuItem<PhysicalKeyboardKey>(
-                value: PhysicalKeyboardKey.mediaPlayPause,
-                child: const Text('Media: Play/Pause'),
-              ),
-              PopupMenuItem<PhysicalKeyboardKey>(
-                value: PhysicalKeyboardKey.mediaStop,
-                child: const Text('Media: Stop'),
-              ),
-              PopupMenuItem<PhysicalKeyboardKey>(
-                value: PhysicalKeyboardKey.mediaTrackPrevious,
-                child: const Text('Media: Previous'),
-              ),
-              PopupMenuItem<PhysicalKeyboardKey>(
-                value: PhysicalKeyboardKey.mediaTrackNext,
-                child: const Text('Media: Next'),
-              ),
-              PopupMenuItem<PhysicalKeyboardKey>(
-                value: PhysicalKeyboardKey.audioVolumeUp,
-                child: const Text('Media: Volume Up'),
-              ),
-              PopupMenuItem<PhysicalKeyboardKey>(
-                value: PhysicalKeyboardKey.audioVolumeDown,
-                child: const Text('Media: Volume Down'),
-              ),
-            ],
-            onSelected: (key) {
-              keyPair.physicalKey = key;
-              keyPair.logicalKey = null;
-
-              setState(() {});
-            },
-            child: ListTile(
-              leading: Icon(Icons.music_note_outlined),
-              trailing: Row(
-                mainAxisSize: MainAxisSize.min,
-                children: [
-                  if (keyPair.isSpecialKey) Checkbox(value: true, onChanged: null),
-                  Icon(Icons.arrow_right),
-                ],
-              ),
-              title: Text('Simulate Media key'),
-            ),
-          ),
-        ),
-    ];
-
     final icon = Container(
       constraints: BoxConstraints(minHeight: iconSize, minWidth: iconSize),
       child: Column(
@@ -287,51 +185,7 @@ class _TouchAreaSetupPageState extends State<TouchAreaSetupPage> {
                 ],
               ),
             ),
-          PopupMenuButton<PhysicalKeyboardKey>(
-            enabled: enableTouch,
-            itemBuilder: (context) => [
-              if (actions.length > 1) ...actions,
-              PopupMenuItem<PhysicalKeyboardKey>(
-                value: null,
-                onTap: () {
-                  keyPair.isLongPress = !keyPair.isLongPress;
-                  setState(() {});
-                },
-                child: CheckboxListTile(
-                  value: keyPair.isLongPress,
-                  onChanged: (value) {
-                    keyPair.isLongPress = value ?? false;
-                    setState(() {});
-                    Navigator.of(context).pop();
-                  },
-                  title: const Text('Long Press Mode (vs. repeating)'),
-                ),
-              ),
-              PopupMenuDivider(),
-              PopupMenuItem<PhysicalKeyboardKey>(
-                value: null,
-                child: ListTile(
-                  title: const Text('Delete Keymap'),
-                  leading: Icon(Icons.delete, color: Colors.red),
-                ),
-                onTap: () {
-                  actionHandler.supportedApp!.keymap.keyPairs.remove(keyPair);
-                  setState(() {});
-                },
-              ),
-            ],
-            onSelected: (key) {
-              keyPair.physicalKey = key;
-              keyPair.logicalKey = null;
-              setState(() {});
-            },
-            child: Row(
-              children: [
-                KeypairExplanation(withKey: true, keyPair: keyPair),
-                Icon(Icons.more_vert),
-              ],
-            ),
-          ),
+          KeypairExplanation(withKey: true, keyPair: keyPair),
         ],
       ),
     );
@@ -417,20 +271,18 @@ class _TouchAreaSetupPageState extends State<TouchAreaSetupPage> {
                     ),
                   ),
 
-                ...?actionHandler.supportedApp?.keymap.keyPairs.map((keyPair) {
-                  return _buildDraggableArea(
-                    enableTouch: true,
-                    keyPair: keyPair,
-                    onPositionChanged: (newPos) {
-                      // convert to percentage
-                      final relativeX = ((newPos.dx - _imageRect.left) / _imageRect.width).clamp(0.0, 1.0);
-                      final relativeY = ((newPos.dy - _imageRect.top) / _imageRect.height).clamp(0.0, 1.0);
-                      keyPair.touchPosition = Offset(relativeX * 100.0, relativeY * 100.0);
-                      setState(() {});
-                    },
-                    color: Colors.red,
-                  );
-                }),
+                _buildDraggableArea(
+                  enableTouch: true,
+                  keyPair: widget.keyPair,
+                  onPositionChanged: (newPos) {
+                    // convert to percentage
+                    final relativeX = ((newPos.dx - _imageRect.left) / _imageRect.width).clamp(0.0, 1.0);
+                    final relativeY = ((newPos.dy - _imageRect.top) / _imageRect.height).clamp(0.0, 1.0);
+                    widget.keyPair.touchPosition = Offset(relativeX * 100.0, relativeY * 100.0);
+                    setState(() {});
+                  },
+                  color: Colors.red,
+                ),
 
                 Positioned.fill(child: Testbed()),
 
@@ -480,6 +332,7 @@ class _TouchAreaSetupPageState extends State<TouchAreaSetupPage> {
                             child: Text('Reset'),
                             onTap: () {
                               _backgroundImage = null;
+
                               actionHandler.supportedApp?.keymap.reset();
                               setState(() {});
                             },
@@ -510,6 +363,7 @@ class KeypairExplanation extends StatelessWidget {
   Widget build(BuildContext context) {
     return Wrap(
       spacing: 4,
+      runSpacing: 4,
       crossAxisAlignment: WrapCrossAlignment.center,
       children: [
         if (withKey)
@@ -518,8 +372,8 @@ class KeypairExplanation extends StatelessWidget {
           )
         else
           Icon(keyPair.icon),
-        if (keyPair.physicalKey != null && actionHandler.supportedModes.contains(SupportedMode.keyboard)) ...[
-          KeyWidget(
+        if (keyPair.isSpecialKey && actionHandler.supportedModes.contains(SupportedMode.media))
+          _KeyWidget(
             label: switch (keyPair.physicalKey) {
               PhysicalKeyboardKey.mediaPlayPause => 'Play/Pause',
               PhysicalKeyboardKey.mediaStop => 'Stop',
@@ -527,16 +381,50 @@ class KeypairExplanation extends StatelessWidget {
               PhysicalKeyboardKey.mediaTrackNext => 'Next',
               PhysicalKeyboardKey.audioVolumeUp => 'Volume Up',
               PhysicalKeyboardKey.audioVolumeDown => 'Volume Down',
-              _ => keyPair.logicalKey?.keyLabel ?? 'Unknown',
+              _ => 'Unknown',
             },
+          )
+        else if (keyPair.physicalKey != null && actionHandler.supportedModes.contains(SupportedMode.keyboard)) ...[
+          _KeyWidget(
+            label: keyPair.logicalKey?.keyLabel ?? 'Unknown',
           ),
           if (keyPair.isLongPress) Text('long\npress', style: TextStyle(fontSize: 10)),
         ] else ...[
-          if (!withKey)
-            KeyWidget(label: 'X: ${keyPair.touchPosition.dx.toInt()}, Y: ${keyPair.touchPosition.dy.toInt()}'),
+          if (!withKey && keyPair.touchPosition != Offset.zero)
+            _KeyWidget(label: 'X:${keyPair.touchPosition.dx.toInt()}, Y:${keyPair.touchPosition.dy.toInt()}'),
           if (keyPair.isLongPress) Text('long\npress', style: TextStyle(fontSize: 10)),
         ],
       ],
+    );
+  }
+}
+
+class _KeyWidget extends StatelessWidget {
+  final String label;
+  const _KeyWidget({super.key, required this.label});
+
+  @override
+  Widget build(BuildContext context) {
+    return IntrinsicWidth(
+      child: Container(
+        padding: const EdgeInsets.symmetric(horizontal: 4, vertical: 4),
+        constraints: BoxConstraints(minWidth: 30),
+        decoration: BoxDecoration(
+          border: Border.all(color: Theme.of(context).colorScheme.primary),
+          borderRadius: BorderRadius.circular(4),
+          color: Theme.of(context).colorScheme.primaryContainer,
+        ),
+        child: Center(
+          child: Text(
+            label.splitByUpperCase(),
+            style: TextStyle(
+              fontFamily: 'monospace',
+              fontSize: 12,
+              color: Theme.of(context).colorScheme.onPrimaryContainer,
+            ),
+          ),
+        ),
+      ),
     );
   }
 }