+++ lib/widgets/keymap_explanation.dart
@@ -1,195 +1,284 @@
+import 'dart:async';
+
 import 'package:dartx/dartx.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
 import 'package:swift_control/main.dart';
 import 'package:swift_control/pages/device.dart';
-import 'package:swift_control/utils/keymap/buttons.dart';
+import 'package:swift_control/utils/actions/base_actions.dart';
+import 'package:swift_control/utils/keymap/apps/custom_app.dart';
 import 'package:swift_control/utils/keymap/keymap.dart';
+import 'package:swift_control/utils/keymap/manager.dart';
+import 'package:swift_control/widgets/button_widget.dart';
+import 'package:swift_control/widgets/custom_keymap_selector.dart';
 
 import '../pages/touch_area.dart';
-import '../utils/actions/base_actions.dart';
 
-class KeymapExplanation extends StatelessWidget {
+class KeymapExplanation extends StatefulWidget {
   final Keymap keymap;
   final VoidCallback onUpdate;
   const KeymapExplanation({super.key, required this.keymap, required this.onUpdate});
 
   @override
+  State<KeymapExplanation> createState() => _KeymapExplanationState();
+}
+
+class _KeymapExplanationState extends State<KeymapExplanation> {
+  late StreamSubscription<void> _updateStreamListener;
+
+  @override
+  void initState() {
+    super.initState();
+    _updateStreamListener = widget.keymap.updateStream.listen((_) {
+      setState(() {});
+    });
+  }
+
+  @override
+  void didUpdateWidget(KeymapExplanation oldWidget) {
+    super.didUpdateWidget(oldWidget);
+    if (oldWidget.keymap != widget.keymap) {
+      _updateStreamListener.cancel();
+      _updateStreamListener = widget.keymap.updateStream.listen((_) {
+        setState(() {});
+      });
+    }
+  }
+
+  @override
+  void dispose() {
+    super.dispose();
+    _updateStreamListener.cancel();
+  }
+
+  @override
   Widget build(BuildContext context) {
-    final connectedDevice = connection.devices.firstOrNull;
-
-    final availableKeypairs = keymap.keyPairs.filter(
-      (e) => connectedDevice?.availableButtons.containsAny(e.buttons) ?? true,
-    );
-
-    final keyboardGroups = availableKeypairs
-        .filter((e) => e.physicalKey != null && actionHandler.supportedModes.contains(SupportedMode.keyboard))
-        .groupBy((element) => '${element.physicalKey?.usbHidUsage}-${element.isLongPress}');
-    final touchGroups = availableKeypairs
-        .filter(
-          (e) =>
-              (e.physicalKey == null || !actionHandler.supportedModes.contains(SupportedMode.keyboard)) &&
-              e.touchPosition != Offset.zero,
-        )
-        .groupBy((element) => '${element.touchPosition.dx}-${element.touchPosition.dy}-${element.isLongPress}');
+    final availableKeypairs = widget.keymap.keyPairs;
+    final allAvailableButtons = connection.devices.flatMap((d) => d.availableButtons);
 
     return Column(
       crossAxisAlignment: CrossAxisAlignment.start,
       mainAxisSize: MainAxisSize.min,
       spacing: 8,
       children: [
-        if (keymap.keyPairs.isEmpty)
-          Text('No key mappings found. Please customize the keymap.')
-        else
-          Table(
-            border: TableBorder.symmetric(
-              borderRadius: BorderRadius.circular(9),
-              inside: BorderSide(
-                color: Theme.of(context).colorScheme.primaryContainer,
-              ),
-              outside: BorderSide(
-                color: Theme.of(context).colorScheme.primaryContainer,
-              ),
+        Table(
+          border: TableBorder.symmetric(
+            borderRadius: BorderRadius.circular(9),
+            inside: BorderSide(
+              color: Theme.of(context).colorScheme.primaryContainer,
             ),
-            children: [
+            outside: BorderSide(
+              color: Theme.of(context).colorScheme.primaryContainer,
+            ),
+          ),
+          children: [
+            TableRow(
+              children: [
+                Padding(
+                  padding: const EdgeInsets.all(6),
+                  child: Text(
+                    'Button on your ${connection.devices.isEmpty ? 'Device' : connection.devices.joinToString(transform: (d) => d.name.screenshot)}',
+                    style: Theme.of(context).textTheme.bodyMedium?.copyWith(fontWeight: FontWeight.bold),
+                  ),
+                ),
+                Padding(
+                  padding: const EdgeInsets.all(6),
+                  child: Text(
+                    'Action',
+                    style: Theme.of(context).textTheme.bodyMedium?.copyWith(fontWeight: FontWeight.bold),
+                  ),
+                ),
+              ],
+            ),
+            for (final keyPair in availableKeypairs) ...[
               TableRow(
                 children: [
                   Padding(
                     padding: const EdgeInsets.all(6),
-                    child: Text(
-                      'Button on your ${connectedDevice?.device.name?.screenshot ?? connectedDevice?.runtimeType ?? 'device'}',
-                      style: Theme.of(context).textTheme.bodyMedium?.copyWith(fontWeight: FontWeight.bold),
+                    child: Wrap(
+                      spacing: 8,
+                      runSpacing: 8,
+                      crossAxisAlignment: WrapCrossAlignment.center,
+                      children: [
+                        if (actionHandler.supportedApp is! CustomApp)
+                          for (final button in keyPair.buttons.filter((b) => allAvailableButtons.contains(b)))
+                            IntrinsicWidth(child: ButtonWidget(button: button))
+                        else
+                          for (final button in keyPair.buttons) IntrinsicWidth(child: ButtonWidget(button: button)),
+                      ],
                     ),
                   ),
                   Padding(
                     padding: const EdgeInsets.all(6),
-                    child: Text(
-                      'Action',
-                      style: Theme.of(context).textTheme.bodyMedium?.copyWith(fontWeight: FontWeight.bold),
-                    ),
+                    child: _ButtonEditor(keyPair: keyPair, onUpdate: widget.onUpdate),
                   ),
                 ],
               ),
-              for (final pair in keyboardGroups.entries) ...[
-                TableRow(
-                  children: [
-                    Padding(
-                      padding: const EdgeInsets.all(6),
-                      child: Wrap(
-                        spacing: 8,
-                        runSpacing: 8,
-                        children: [
-                          for (final keyPair in pair.value)
-                            for (final button in keyPair.buttons)
-                              if (connectedDevice?.availableButtons.contains(button) ?? true)
-                                IntrinsicWidth(child: ButtonWidget(button: button)),
-                        ],
-                      ),
-                    ),
-                    Padding(
-                      padding: const EdgeInsets.all(6),
-                      child: KeypairExplanation(keyPair: pair.value.first),
-                    ),
-                  ],
-                ),
-              ],
-              for (final pair in touchGroups.entries) ...[
-                TableRow(
-                  children: [
-                    Padding(
-                      padding: const EdgeInsets.all(6),
-                      child: Wrap(
-                        spacing: 8,
-                        runSpacing: 8,
-                        children: [
-                          for (final keyPair in pair.value)
-                            for (final button in keyPair.buttons)
-                              if (connectedDevice?.availableButtons.contains(button) ?? true)
-                                ButtonWidget(button: button),
-                        ],
-                      ),
-                    ),
-                    Padding(
-                      padding: const EdgeInsets.all(6),
-                      child: KeypairExplanation(keyPair: pair.value.first),
-                    ),
-                  ],
-                ),
-              ],
             ],
-          ),
+          ],
+        ),
       ],
     );
   }
 }
 
-class KeyWidget extends StatelessWidget {
-  final String label;
-  const KeyWidget({super.key, required this.label});
+class _ButtonEditor extends StatelessWidget {
+  final KeyPair keyPair;
+  final VoidCallback onUpdate;
+  const _ButtonEditor({required this.onUpdate, super.key, required this.keyPair});
 
   @override
   Widget build(BuildContext context) {
-    return IntrinsicWidth(
-      child: Container(
-        padding: const EdgeInsets.symmetric(horizontal: 4, vertical: 4),
-        constraints: BoxConstraints(minWidth: 30),
-        decoration: BoxDecoration(
-          border: Border.all(color: Theme.of(context).colorScheme.primary),
-          borderRadius: BorderRadius.circular(4),
-          color: Theme.of(context).colorScheme.primaryContainer,
+    final actions = [
+      if (actionHandler.supportedModes.contains(SupportedMode.keyboard))
+        PopupMenuItem<PhysicalKeyboardKey>(
+          value: null,
+          child: ListTile(
+            leading: Icon(Icons.keyboard_alt_outlined),
+            title: const Text('Simulate Keyboard shortcut'),
+            trailing: keyPair.physicalKey != null ? Checkbox(value: true, onChanged: null) : null,
+          ),
+          onTap: () async {
+            await showDialog<void>(
+              context: context,
+              barrierDismissible: false, // enable Escape key
+              builder: (c) =>
+                  HotKeyListenerDialog(customApp: actionHandler.supportedApp! as CustomApp, keyPair: keyPair),
+            );
+            onUpdate();
+          },
         ),
-        child: Center(
-          child: Text(
-            label.splitByUpperCase(),
-            style: TextStyle(
-              fontFamily: 'monospace',
-              fontSize: 12,
-              color: Theme.of(context).colorScheme.onPrimaryContainer,
+      if (actionHandler.supportedModes.contains(SupportedMode.touch))
+        PopupMenuItem<PhysicalKeyboardKey>(
+          value: null,
+          child: ListTile(
+            title: const Text('Simulate Touch'),
+            leading: Icon(Icons.touch_app_outlined),
+            trailing: keyPair.physicalKey == null && keyPair.touchPosition != Offset.zero
+                ? Checkbox(value: true, onChanged: null)
+                : null,
+          ),
+          onTap: () async {
+            if (keyPair.touchPosition == Offset.zero) {
+              keyPair.touchPosition = Offset(50, 50);
+            }
+            keyPair.physicalKey = null;
+            keyPair.logicalKey = null;
+            await Navigator.of(context).push<bool?>(
+              MaterialPageRoute(
+                builder: (c) => TouchAreaSetupPage(
+                  keyPair: keyPair,
+                ),
+              ),
+            );
+            onUpdate();
+          },
+        ),
+
+      if (actionHandler.supportedModes.contains(SupportedMode.media))
+        PopupMenuItem<PhysicalKeyboardKey>(
+          child: PopupMenuButton<PhysicalKeyboardKey>(
+            padding: EdgeInsets.zero,
+            itemBuilder: (context) => [
+              PopupMenuItem<PhysicalKeyboardKey>(
+                value: PhysicalKeyboardKey.mediaPlayPause,
+                child: const Text('Media: Play/Pause'),
+              ),
+              PopupMenuItem<PhysicalKeyboardKey>(
+                value: PhysicalKeyboardKey.mediaStop,
+                child: const Text('Media: Stop'),
+              ),
+              PopupMenuItem<PhysicalKeyboardKey>(
+                value: PhysicalKeyboardKey.mediaTrackPrevious,
+                child: const Text('Media: Previous'),
+              ),
+              PopupMenuItem<PhysicalKeyboardKey>(
+                value: PhysicalKeyboardKey.mediaTrackNext,
+                child: const Text('Media: Next'),
+              ),
+              PopupMenuItem<PhysicalKeyboardKey>(
+                value: PhysicalKeyboardKey.audioVolumeUp,
+                child: const Text('Media: Volume Up'),
+              ),
+              PopupMenuItem<PhysicalKeyboardKey>(
+                value: PhysicalKeyboardKey.audioVolumeDown,
+                child: const Text('Media: Volume Down'),
+              ),
+            ],
+            onSelected: (key) {
+              keyPair.physicalKey = key;
+              keyPair.logicalKey = null;
+
+              onUpdate();
+            },
+            child: ListTile(
+              leading: Icon(Icons.music_note_outlined),
+              trailing: Row(
+                mainAxisSize: MainAxisSize.min,
+                children: [
+                  if (keyPair.isSpecialKey) Checkbox(value: true, onChanged: null),
+                  Icon(Icons.arrow_right),
+                ],
+              ),
+              title: Text('Simulate Media key'),
             ),
           ),
         ),
-      ),
-    );
-  }
-}
+    ];
 
-class ButtonWidget extends StatelessWidget {
-  final ControllerButton button;
-  final bool big;
-  const ButtonWidget({super.key, required this.button, this.big = false});
+    return Row(
+      mainAxisAlignment: MainAxisAlignment.spaceBetween,
+      children: [
+        if (keyPair.buttons.isNotEmpty && (keyPair.physicalKey != null || keyPair.touchPosition != Offset.zero))
+          Expanded(
+            child: KeypairExplanation(
+              keyPair: keyPair,
+            ),
+          )
+        else
+          Text('No action assigned'),
 
-  @override
-  Widget build(BuildContext context) {
-    return IntrinsicWidth(
-      child: Container(
-        padding: const EdgeInsets.symmetric(horizontal: 4, vertical: 4),
-        constraints: BoxConstraints(
-          minWidth: big && button.color != null ? 40 : 30,
-          minHeight: big && button.color != null ? 40 : 0,
-        ),
-        decoration: BoxDecoration(
-          border: Border.all(color: button.color != null ? Colors.black : Theme.of(context).colorScheme.primary),
-          shape: button.color != null || button.icon != null ? BoxShape.circle : BoxShape.rectangle,
-          borderRadius: button.color != null || button.icon != null ? null : BorderRadius.circular(4),
-          color: button.color ?? Theme.of(context).colorScheme.primaryContainer,
-        ),
-        child: Center(
-          child: button.icon != null
-              ? Icon(
-                  button.icon,
-                  color: Colors.white,
-                  size: big && button.color != null ? null : 14,
-                )
-              : Text(
-                  button.name.splitByUpperCase(),
-                  style: TextStyle(
-                    fontFamily: 'monospace',
-                    fontSize: big && button.color != null ? 20 : 12,
-                    fontWeight: button.color != null ? FontWeight.bold : null,
-                    color: button.color != null ? Colors.white : Theme.of(context).colorScheme.onPrimaryContainer,
-                  ),
+        if (actionHandler.supportedApp is CustomApp)
+          PopupMenuButton<PhysicalKeyboardKey>(
+            enabled: true,
+            itemBuilder: (context) => [
+              if (actions.length > 1) ...actions,
+              PopupMenuItem<PhysicalKeyboardKey>(
+                value: null,
+                onTap: () {
+                  keyPair.isLongPress = !keyPair.isLongPress;
+                  onUpdate();
+                },
+                child: CheckboxListTile(
+                  value: keyPair.isLongPress,
+                  onChanged: (value) {
+                    keyPair.isLongPress = value ?? false;
+
+                    onUpdate();
+                    Navigator.of(context).pop();
+                  },
+                  title: const Text('Long Press Mode (vs. repeating)'),
                 ),
-        ),
-      ),
+              ),
+            ],
+            onSelected: (key) {
+              keyPair.physicalKey = key;
+              keyPair.logicalKey = null;
+
+              onUpdate();
+            },
+            icon: Icon(Icons.edit),
+          )
+        else
+          IconButton(
+            onPressed: () async {
+              final currentProfile = actionHandler.supportedApp!.name;
+              await KeymapManager().duplicate(context, currentProfile);
+              onUpdate();
+            },
+            icon: Icon(Icons.edit),
+          ),
+      ],
     );
   }
 }